refactor: 검증 위치 변경
