+++ src/main/java/ladder/domain/ladder/Ladder.java
@@ -1,10 +1,15 @@
 package ladder.domain.ladder;
 
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 public class Ladder {
+    private static final String NOT_NUMBER = "\\D";
+    private static final Pattern CHARACTER_SET_NOT_NUMBER = Pattern.compile(NOT_NUMBER);
+
     private final List<Line> lines;
 
     public Ladder(BarGenerator barGenerator, int ladderHeight, int peopleSize) {
@@ -25,5 +30,12 @@ public class Ladder {
 
     public List<Line> getLines() {
         return this.lines;
+    }
+
+    public static void validateNonNumber(String ladderHeight) {
+        Matcher matcher = CHARACTER_SET_NOT_NUMBER.matcher(ladderHeight);
+        if (matcher.find()) {
+            throw new IllegalArgumentException("숫자가 아닌 값은 입력할 수 없습니다.");
+        }
     }
 }
