+++ src/main/java/ladder/view/util/InputValidator.java
@@ -1,12 +1,8 @@
 package ladder.view.util;
 
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class InputValidator {
-    private static final String NOT_NUMBER = "\\D";
-    private static final Pattern CHARACTER_SET_NOT_NUMBER = Pattern.compile(NOT_NUMBER);
 
     private InputValidator() {
     }
@@ -17,18 +13,10 @@ public class InputValidator {
         }
     }
 
-    public static void validateNonNumber(String ladderHeight) {
-        Matcher matcher = CHARACTER_SET_NOT_NUMBER.matcher(ladderHeight);
-        if (matcher.find()) {
-            throw new IllegalArgumentException("숫자가 아닌 값은 입력할 수 없습니다.");
-        }
-    }
-
     public static int validateLadderHeightRange(int ladderHeight) {
         if (ladderHeight < 1 || ladderHeight > 100) {
             throw new IllegalArgumentException("사다리 높이는 1이상 100 이하만 입력할 수 있습니다.");
         }
         return ladderHeight;
     }
-
 }
